API refactor: rename createLoguxStore to createLoguxCreator
